TS-migration: VirtualizedListItems
